Moved `NeroTool` to `joe.tools.nero`.
