+++ lib/src/main/java/com/wjduquette/joe/app/App.java
@@ -1,6 +1,7 @@
 package com.wjduquette.joe.app;
 
 import com.wjduquette.joe.tools.doc.DocTool;
+import com.wjduquette.joe.tools.nero.NeroTool;
 import com.wjduquette.joe.tools.win.WinTool;
 import com.wjduquette.joe.tools.test.TestTool;
 import com.wjduquette.joe.tools.ToolLauncher;
